change int to Long, delete action

+++ src/main/java/ru/balyanova/productData/Product.java
@@ -2,17 +2,17 @@ package ru.balyanova.productData;
 
 public class Product {
 
-    private int id;
+    private Long id;
     private String title;
     private double cost;
 
-    public Product(int id, String title, double cost) {
+    public Product(Long id, String title, double cost) {
         this.id = id;
         this.title = title;
         this.cost = cost;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
